Add a settings endpoint
