+++ app/controllers/app_controller.rb
@@ -1,20 +1,9 @@
 class AppController < ApplicationController
   def index
     if user_signed_in?
-      @settings = {
-        defaultSplitId: default_split_id,
-        admin: current_user.admin?
-      }.to_json
+      # just render the app
     else
       redirect_to new_user_session_path
     end
   end
-
-  private 
-
-  def default_split_id
-    split = Split.find_by_name(EasySettings['default_split'])
-    split ? split.id : nil
-  end
 end
-
