+++ config/routes.rb
@@ -31,6 +31,7 @@ Rails.application.routes.draw do
         delete 'session', to: 'session#destroy'
         resources :data, only: [:index]
         resources :bank_accounts, only: [:index]
+        resources :settings, only: %i(index)
         resources :bets
         resources :upcoming_games, only: [:index]
         resources :odds, only: %i(create show)
